some refactor and display title input field only when component is selected
