+++ frontend/packages/ux-editor/src/components/Properties/EditSubformTableColumns/ColumnElement/ColumnElement.tsx
@@ -9,7 +9,7 @@ import { useStudioEnvironmentParams } from 'app-shared/hooks/useStudioEnvironmen
 import { textResourceByLanguageAndIdSelector } from '../../../../selectors/textResourceSelectors';
 
 export type ColumnElementProps = {
-  layoutSetName: string;
+  subformLayout: string;
   tableColumn: TableColumn;
   columnNumber: number;
   onDeleteColumn: () => void;
@@ -21,7 +21,7 @@ export const ColumnElement = ({
   columnNumber,
   onDeleteColumn,
   onEdit,
-  layoutSetName,
+  subformLayout,
 }: ColumnElementProps): ReactElement => {
   const { t } = useTranslation();
   const [editing, setEditing] = useState(false);
@@ -36,7 +36,7 @@ export const ColumnElement = ({
   if (editing) {
     return (
       <EditColumnElement
-        layoutSetName={layoutSetName}
+        subformLayout={subformLayout}
         sourceColumn={tableColumn}
         columnNumber={columnNumber}
         onDeleteColumn={onDeleteColumn}
